Rename deaths and merge code and update Makefile
