--- test_img/readme.md
@@ -1,1 +1,0 @@
-empty text file to create directory needed for test images
